requirements.txt
flask
pymupdf
python-dotenv
requests
gunicorn
